Scanpath similarity helper functions
arranged into new file: scanpathhelper.py
