+++ PyEyeSim/__init__.py
@@ -20,11 +20,12 @@ import hmmlearn.hmm  as hmm
 from matplotlib.patches import Ellipse
 import platform
 #%%
-from .visualhelper import VisBinnedProg
+from .visualhelper import VisBinnedProg,PlotDurProg,JointBinnedPlot,MeanPlot
+from .scanpathsimhelper import AOIbounds,CreatAoiRects,Rect,SaccadeLine,CalcSim
 
 
 class EyeData:
-    from ._visuals import VisScanPath,MySaccadeVis,VisLOOHMM,VisHMM
+    from ._visuals import VisScanPath,MySaccadeVis,VisLOOHMM,VisHMM,MyTrainTestVis,MySaccadeVis
 
     def __init__(self, name, design,data,x_size,y_size,fixdata=1):
         ''' initalizing eye data object:
@@ -967,36 +968,7 @@ class EyeData:
             self.MyTrainTestVis(DatTr,DatTest,lenTrain,lenTest,totest)
         return DatTr,DatTest,lenTrain,lenTest   
     
-    def MyTrainTestVis(self, DatTr,DatTest,lenTrain,lenTest,totest=0):    
-        ''' make figure for training test - set visualization'''
-        fig,ax=plt.subplots(ncols=2,figsize=(8,3.5))
-        self.MySaccadeVis(ax[0],DatTr,lenTrain,title='training data',alpha=.3)
-        if type(totest)==int:
-            titStr=''
-        else:
-            if len(lenTest)==1:
-                titStr='subj '+str(int(self.suseHMM[totest][0]))
-            else:
-                titStr='n subj '+str(len(totest))
-
-        self.MySaccadeVis(ax[1],DatTest,lenTest,title='test data '+titStr)
-        return 
-          
-    def MySaccadeVis(self,ax,XYdat,lengths,title='',alpha=1):
-        ''' saccade visualization, on input ax, based on combined data 2d array, and lengths 1d array'''
-        ax.set_title(title)
-        ax.set_xlim([0,self.x_size])
-        ax.set_ylim([self.y_size,0])
-        ax.scatter(XYdat[:,0],XYdat[:,1],c=np.arange(len(XYdat[:,0])),alpha=alpha)
-        Idxs=np.cumsum(lengths)
-        for ci in range(len(lengths)):
-            if ci==0:
-                start=0
-            else:
-                start=Idxs[ci-1]
-            ax.plot(XYdat[start:Idxs[ci],0],XYdat[start:Idxs[ci],1],alpha=alpha)
-        return
-            
+    
 
     def FitLOOHMM(self,ncomp,stim,covar='full'):
         ''' fit HMM, N subject times, leaving out once each time
@@ -1280,16 +1252,7 @@ def FitScoreHMMGauss(ncomp,xx,xxt,lenx,lenxxt,covar='full'):
     scte=HMM.score(xxt,lenxxt)/np.sum(lenxxt)
     return HMM,sctr,scte
 
-def MeanPlot(N,Y,yLab=0,xtickL=0,color='darkred',label=None,ax=0):
-    ''' expects data format: row- subjects column-stimuli '''
-    if type(ax)==int:
-        fig,ax=plt.subplots(figsize=(N/2,5))
-    ax.errorbar(np.arange(N),np.nanmean(Y,0),stats.sem(Y,0,nan_policy="omit")*2,linestyle='None',marker='o',color=color,label=label)
-    if type(xtickL)!=int:
-        ax.set_xticks(np.arange(N),labels=xtickL,fontsize=9,rotation=60)
-    ax.set_xlabel('Stimulus',fontsize=14)
-    ax.set_ylabel(yLab,fontsize=14)
-    return None
+
 
 def HistPlot(Y,xtickL=0,ax=0):
     ''' expected data format: row-  subjects column- stimuli '''
@@ -1301,10 +1264,6 @@ def HistPlot(Y,xtickL=0,ax=0):
     ax.set_ylabel('Num observers',fontsize=13)
     return None
 
-
-def AOIbounds(start,end,nDiv):  
-    ''' calcuale AOI bounds, linearly spaced from: start to: end, for nDiv number of divisions'''
-    return np.linspace(start,end,nDiv+1)  
 
 
 def SaliencyMapFilt(Fixies,SD=25,Ind=0):
@@ -1357,32 +1316,6 @@ def StatEntropy(StatP):
     return -np.sum(StatP*LogP)
 
 
-def PlotDurProg(nmax,Y,error,label=''):
-    plt.fill_between(np.arange(1,nmax+1),Y-error,Y+error,alpha=.5)
-    plt.plot(np.arange(1,nmax+1),Y,label=label)
-    plt.xticks(np.intp(np.linspace(1,nmax,5)),np.intp(np.linspace(1,nmax,5)))
-    plt.xlabel('fixation number',fontsize=14)
-    plt.ylabel('fixation duration (ms)',fontsize=14)
-    return 
-
-
-
-
-def JointBinnedPlot(bins,y1,y2,col1='olive',col2='orange',ylabel1='',ylabel2=''):
-    fig,ax1=plt.subplots(figsize=(5,3))
-    ax1.errorbar((bins[0:-1]+bins[1:])/2,np.nanmean(y1,0),stats.sem(y1,0,nan_policy='omit'),color=col1,linewidth=2)
-    ax1.set_xlabel('time (ms)')
-    ax1.yaxis.set_major_locator(ticker.MaxNLocator(5))
-    ax1.set_ylabel(ylabel1,color=col1)
-    ax1.tick_params(axis='y', labelcolor=col1)
-
-    ax2 = ax1.twinx() 
-    ax2.set_ylabel(ylabel2,color=col2)
-    ax2.errorbar((bins[0:-1]+bins[1:])/2,np.nanmean(y2,0),stats.sem(y2,0,nan_policy='omit'),color=col2,linewidth=2)
-    ax2.tick_params(axis='y', labelcolor=col2)
-    ax2.yaxis.set_major_locator(ticker.MaxNLocator(5))
-
-    return ax1,ax2
 
 
 
@@ -1407,106 +1340,6 @@ def draw_ellipse(position, covariance, ax=None, **kwargs):
         ax.add_patch(Ellipse(position, nsig * width, nsig * height,
                              angle, **kwargs))
         
-
-
-def CreatAoiRects(nHorD,nVerD,BoundsX,BoundsY):
-    AOIRects=[]
-    for p in range( np.shape(BoundsX)[0]):
-        AOIboundsH=AOIbounds(BoundsX[p,0],BoundsX[p,1],nHorD)
-        AOIboundsV=AOIbounds(BoundsY[p,0],BoundsY[p,1],nVerD)
-        print(AOIboundsH)
-        print(AOIboundsV)
-
-        AOIRects.append([])
-        for h in range(nHorD):
-            AOIRects[p].append([])
-            for v in range(nVerD):
-                AOIRects[p][h].append(Rect(AOIboundsH[h],AOIboundsV[v],AOIboundsH[h+1],AOIboundsV[v+1]))  # store AOIs as objects
-    return AOIRects
-
-
-
-class Rect:
-     def __init__(self, x1,y1,x2,y2):
-        if x1 > x2: 
-            x1,x2=x2,x1 # to start with smaller x
-        if y1 > y2:
-            y1,y2,=y2,y1 # to start with smaller y
-        self.x1=x1
-        self.y1=y1
-        self.x2=x2
-        self.y2=y2    
-     def Vis(self,alp=.2,Col='b'):
-        plt.plot([self.x1,self.x1],[self.y1,self.y2],alpha=alp,color=Col)
-        plt.plot([self.x2,self.x2],[self.y1,self.y2],alpha=alp,color=Col)
-        plt.plot([self.x1,self.x2],[self.y1,self.y1],alpha=alp,color=Col)
-        plt.plot([self.x1,self.x2],[self.y2,self.y2],alpha=alp,color=Col)
-        
-     def Contains(self,x,y):
-        if x>self.x1 and x<self.x2 and y > self.y1 and y < self.y2:              
-            return True
-        else:
-            return False
-        
-     def Cross(self,LinePoints):
-         CrossXidx=np.nonzero((LinePoints[0]>self.x1)&(LinePoints[0]<self.x2))
-         if len(CrossXidx)>0:
-             if any(LinePoints[1][CrossXidx]>self.y1)==True and any(LinePoints[1][CrossXidx]<self.y2)==True:
-                 return True
-             else:
-                 return False
-         else:
-            return False
-                     
-     def Contains2(self,LinePoints):
-      
-         CrossXidx=np.nonzero((LinePoints[:,0]>self.x1)&(LinePoints[:,0]<self.x2))
-         CrossYidx=np.nonzero((LinePoints[:,1]>self.y1)&(LinePoints[:,1]<self.y2))
-         return  np.intersect1d(CrossXidx,CrossYidx)    
-     
-        
-
-class SaccadeLine:
-    def __init__(self, coordvect):  # create a saccade object with start and end pixels
-        self.x1=coordvect[0]
-        self.y1=coordvect[1]
-        self.x2=coordvect[2]
-        self.y2=coordvect[3]
-    def Coords(self):   
-        return self.x1,self.y1,self.x2,self.y2
-    def length(self):   # length of saccade
-        return int(np.sqrt((self.x2-self.x1)**2+(self.y2-self.y1)**2))
-    
-    def lengthHor(self):  # horizontal length of saccade
-        return  np.abs(self.x2-self.x1)
-    
-    def lengthVer(self):  # vertical length of saccade
-        return  np.abs(self.y2-self.y1)
-    
-    def Angle(self):   # angle of saccade (0-360 deg)
-        Ang=np.degrees(np.arccos((self.x2-self.x1)/self.length()))  #calculate angel of saccades
-        if self.y2 < self.y1:  # if downward saccade
-            Ang=360-Ang  
-        return Ang
-    def Vis(self,alp=.2,Col='k'):  # plot saccade
-        plt.plot([self.x1,self.x2],[self.y1,self.y2],alpha=alp,color=Col)
-        return
-    
-    def LinePoints(self):  # use dots with density of 1dot/1pixel to approximate line.
-        LineX=np.linspace(self.x1,self.x2,self.length())
-        LineY=np.linspace(self.y1,self.y2,self.length())
-        return LineX,LineY
-    
-
-def CalcSim(saccades1,saccades2,Thr=5):
-    ''' calculcate angle based similarity for two arrays of saccade objects (for each cell)'''
-    A=matlib.repmat(saccades1,len(saccades2),1)   # matrix of s1 saccades in cell
-    B=matlib.repmat(saccades2,len(saccades1),1).T  # matrix of s2 saccades in cell
-    simsacn=np.sum(np.abs(A-B)<Thr)
-    A[A>180]-=180
-    A[A<180]+=180                                       
-    simsacn+=np.sum(np.abs(A-B)<Thr) 
-    return simsacn
